requirements.txt
openai
langchain
pymupdf
python-dotenv
faiss-cpu
fastapi
uvicorn[standard]
chainlit
loguru
sentry-sdk